requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf_yasg
coreapi 
pyyaml